@@ -1,21 +1,28 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 apply from: 'constants.gradle'
 
-def keystorePropertiesFile = rootProject.file("app/keystore.properties")
+def keystorePropertiesFile = rootProject.file("keystore.properties")
 def keystoreProperties = new Properties()
 keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
 
 buildscript {
+    ext.kotlinVersion = '1.1.60'
     repositories {
         jcenter()
+        google()
+        maven { url 'https://maven.fabric.io/public' }
+        maven { url 'https://oss.sonatype.org/content/repositories/snapshots/' }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:2.3.3'
+        classpath 'com.android.tools.build:gradle:3.0.1'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
-        classpath 'com.google.gms:google-services:3.0.0'
-        classpath 'com.jakewharton:butterknife-gradle-plugin:8.8.1'
+        classpath 'com.google.gms:google-services:3.1.1'
+        classpath 'com.jakewharton:butterknife-gradle-plugin:9.0.0-SNAPSHOT'
+        classpath 'io.fabric.tools:gradle:1.24.4'
+        classpath 'com.google.firebase:firebase-plugins:1.1.1'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
     }
 }
 
@@ -23,9 +30,7 @@ allprojects {
     repositories {
         jcenter()
         google()
-        maven {
-            url 'https://oss.sonatype.org/content/repositories/snapshots/'
-        }
+        maven { url 'https://oss.sonatype.org/content/repositories/snapshots/' }
     }
 }
 
@@ -41,16 +46,16 @@ subprojects {
                 javaMaxHeapSize "2g"
             }
             compileOptions {
-                sourceCompatibility JavaVersion.VERSION_1_7
-                targetCompatibility JavaVersion.VERSION_1_7
+                sourceCompatibility JavaVersion.VERSION_1_8
+                targetCompatibility JavaVersion.VERSION_1_8
             }
             if (isAndroidProject) {
                 signingConfigs {
                     release {
+                        storeFile file(rootProject.file("nas2skupa.jks"))
+                        storePassword keystoreProperties['storePassword']
                         keyAlias keystoreProperties['keyAlias']
                         keyPassword keystoreProperties['keyPassword']
-                        storeFile file(keystoreProperties['storeFile'])
-                        storePassword keystoreProperties['storePassword']
                     }
                 }
                 buildTypes {
@@ -65,7 +70,7 @@ subprojects {
         }
         dependencies {
             annotationProcessor "org.androidannotations:androidannotations:$androidAnnotationsVersion"
-            compile "org.androidannotations:androidannotations-api:$androidAnnotationsVersion"
+            implementation "org.androidannotations:androidannotations-api:$androidAnnotationsVersion"
         }
     }
 }
