@@ -2,6 +2,7 @@ package hr.nas2skupa.eleventhhour.admin;
 
 
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.ActivityOptionsCompat;
 import android.support.v4.app.Fragment;
@@ -9,10 +10,13 @@ import android.support.v4.util.Pair;
 import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
 import android.view.View;
+import android.view.ViewGroup;
 import android.view.animation.OvershootInterpolator;
 
 import com.firebase.ui.database.FirebaseRecyclerAdapter;
+import com.firebase.ui.database.FirebaseRecyclerOptions;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.Query;
@@ -20,14 +24,13 @@ import com.google.firebase.database.Query;
 import org.androidannotations.annotations.EFragment;
 import org.androidannotations.annotations.ViewById;
 
+import hr.nas2skupa.eleventhhour.admin.viewholders.CategoryViewHolder;
 import hr.nas2skupa.eleventhhour.common.model.Category;
 import jp.wasabeef.recyclerview.animators.LandingAnimator;
 
 @EFragment(R.layout.fragment_categories)
 public class CategoriesFragment extends Fragment {
     @ViewById RecyclerView recyclerView;
-
-    private FirebaseRecyclerAdapter<Category, CategoryViewHolder> adapter;
 
     public CategoriesFragment() {
         // Required empty public constructor
@@ -45,30 +48,26 @@ public class CategoriesFragment extends Fragment {
 
         DatabaseReference database = FirebaseDatabase.getInstance().getReference();
         Query query = database.child("app/categories");
-        adapter = new FirebaseRecyclerAdapter<Category, CategoryViewHolder>(Category.class, R.layout.item_category
-                , CategoryViewHolder.class, query) {
+        FirebaseRecyclerOptions<Category> options = new FirebaseRecyclerOptions.Builder<Category>()
+                .setQuery(query, Category.class)
+                .setLifecycleOwner(this)
+                .build();
+        FirebaseRecyclerAdapter<Category, CategoryViewHolder> adapter = new FirebaseRecyclerAdapter<Category, CategoryViewHolder>(options) {
             @Override
-            protected void populateViewHolder(final CategoryViewHolder viewHolder, final Category model, int position) {
+            public CategoryViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+                return new CategoryViewHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.item_category, parent, false));
+            }
+
+            @Override
+            protected void onBindViewHolder(@NonNull final CategoryViewHolder viewHolder, int position, @NonNull Category model) {
                 final DatabaseReference categoryRef = getRef(position);
                 final String categoryKey = categoryRef.getKey();
 
                 viewHolder.bindToCategory(model);
-                viewHolder.itemView.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View view) {
-                        startCategoryActivity(viewHolder.itemView, categoryKey);
-                    }
-                });
+                viewHolder.itemView.setOnClickListener(view -> startCategoryActivity(viewHolder.itemView, categoryKey));
             }
         };
         recyclerView.setAdapter(adapter);
-    }
-
-    @Override
-    public void onDestroy() {
-        super.onDestroy();
-
-        if (adapter != null) adapter.cleanup();
     }
 
     private void startCategoryActivity(View card, String categoryKey) {
